% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute the Manhattan Norm Distance Between Two Graphs}
\usage{
compute_distance(G, M)
}
\arguments{
\item{G}{A square adjacency matrix representing a brain network.}

\item{M}{A square adjacency matrix representing the central brain network.}
}
\value{
A numeric value representing the Manhattan distance between G and M.
}
\description{
This function calculates the Manhattan norm (L1 norm) distance between two
adjacency matrices representing brain networks.
}
\examples{
# Generate synthetic data
G1 <- generate_random_graph(n_nodes = 5, edge_prob = 0.1)
G2 <- generate_random_graph(n_nodes = 5, edge_prob = 0.1)
central_graph <- compute_central_graph(list(G1, G2))
distance <- compute_distance(G1, central_graph)
}
