% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_graph.R
\name{generate_random_graph}
\alias{generate_random_graph}
\title{Generate a Random Symmetric Adjacency Matrix}
\usage{
generate_random_graph(n_nodes, edge_prob = 0.1)
}
\arguments{
\item{n_nodes}{An integer specifying the number of nodes (brain regions).}

\item{edge_prob}{A numeric value between 0 and 1 specifying the probability
of an edge existing between any two nodes.}
}
\value{
A symmetric binary adjacency matrix with no self-loops.
}
\description{
This function generates a random symmetric adjacency matrix representing
a brain network. The adjacency matrix is binary, with edges present based
on a specified probability.
}
\examples{
G <- generate_random_graph(n_nodes = 10, edge_prob = 0.1)
}
